Update map.rs re-structure wrong directories and update player movement

+++ OhMaze/ohmaze/src/entities/item.rs
@@ -1,4 +1,4 @@
 pub enum Item {
     Sword, // Wtf
-    Key(u8),
+    Key(usize),
 }
